service_generator/requirements.txt
google-cloud-pubsub
google-cloud-pubsub
flask
